Isolate import item processing away from the network layer
